feat: create basic documentation page and PoC next-rest-framework implementation
The more I work with supabase, the more obvious it becomes that a
middleman between the DB and end user will be necessary. This project is
intended to be this middleman.

The
[next-rest-framework](https://github.com/blomqma/next-rest-framework)
seems to provide a nice API for generating OpenAPI documentation for API
endpoints and vercel should be a good fit for hosting.

Note that the Swagger UI exposed by next-rest-framework is not yet
advanced enough for our requirements. For this reason, this commit
implements the Swagger UI directly, without the next-rest-framework
middleman.