feat: configure backing database
The project will use the [Prisma](https://www.prisma.io/) ORM to
configure and use its database. The database itself will be
[Vercel's postgres](https://vercel.com/docs/storage/vercel-postgres)
offering.

This commit configures the devcontainer with an associated postgres
instance for use as a development environment, prisma and other basics
to support a database in the project.