style: suffix prisma entities with `Entity`
A prisma `Profile` and our abstract `Profile` concept do not overlap
properly (this issue is not limited to profiles!). For this reason,
we'll need to introduce a custom `ProfileModel` which matches our
expectations.

To prevent confusion, the idea is for database entities to have the
suffix `Entity`, DTOs to have the suffix `DTO` and our abstract
representation to have the suffix `Model`. I.e.:
* `ProfileDTO` - used to transfer data over the API
* `ProfileModel` - used to work with data in code
* `ProfileEntity` - used to store data in the database.