fix: run `prisma generate` as part of `postinstall` script
This fixes an issue noted by the following logprint on vercel with the
fix recommended by the provided link:
```
Prisma has detected that this project was built on Vercel, which caches
dependencies. This leads to an outdated Prisma Client because Prisma's
auto-generation isn't triggered. To fix this, make sure to run the
`prisma generate` command during the build process.

Learn how: https://pris.ly/d/vercel-build
```