build: use envsafe to improve environment variable typing
envsafe allows us to avoid `undefined` environment variables and also
raises build-time warnings if a variable is missing. Further, it makes
it easier to introduce new developers to the project with pre-configured
`devDefault` values for some variables.

One drawback of using `envsafe` is that JS code must be run before the
environment variables are usable. This caused some issues with e.g.
prisma, which relies on an external rust tool. I managed to work around
this by implementing a custom "dotenv-cli" tool - `load_envvars.ts` -
loosely based on
https://github.com/entropitor/dotenv-cli/blob/master/cli.js