feat: support marking venues as permanently closed
I realized we run into a bit of an issue if venues are closed. On the
one hand, we do not want to list them anymore. On the other, there may
have been past events organized at the venue.

This commit attempts to solve this issue by allowing venues to be marked
"permanently closed" as opposed to being deleted.