refactor: make the core api handler more modular with middleware
This makes it easier to extend it with more functionality and
improves readability.

Built-in parameter parsing is removed in this commit. I found the
concept a bit more complicated than initially intended and have decided
to wait a bit with the implementation until I have actual endpoints that
use parameters.