feat: allow deletion of entities with connected images
Initially, the idea was that by restricting deletions of entities with
connected images, we would be forced to clean up any dangling images at
the same time. However, in practice, it seems much easier to run an
automated cleanup every now and then.

This commit allows entities that are connected to images to be deleted
but does not implement an automated cleanup - this will need to be done
later.