fix: enforce image variant at schema level
The old database schema did not structurally restrict the number of
images - or the number of images with a specific $VARIANT - connected
to another entity. While this constraint could probably be enforced
logically (i.e. by us remembering not to change the variant of an
uploaded image or running programmatic checks), it would be better
with structural enforcement (i.e. making it so that the database
schema does not allow us to make mistakes).

This commit solves the problem by removing the `${TABLE}ImageJunction`
tables and adding `${VARIANT}ImageId` fields in the `$TABLE` table. This
ensures that we can never add more images than are supported by the type
as well as that we are never able to add more than 1 image per
`$VARIANT`.