fix: make it so that clients do not need to track the cloudflare id
Originally, the idea was to only create our local image reference once
the image was properly uploaded to cloudflare. However, this means
clients would need to refer to the cloudflare id between requesting an
upload url and actually uploading the image, which is undesirable.

This commit instead makes it so that clients only need to track our
image id, at the cost of some uncertainty regarding the image's
availability on our side. However, we already had a similar availability
issue with images that are deleted on cloudflare but not in our
database, meaning that we would have needed to figure out a fix for
"dangling" images anyways.

Note that fixing "dangling" image references (i.e. local references to
cloudflare images that do not exist) are not done in this task, but
rather split off into
https://github.com/dansdata-se/api/issues/2