style(eslint): enable `reportUnusedDisableDirectives`
I have not found a way of making this an error, but at least we'll get a
warning when unused `eslint-disable-*` are in use.

I would prefer this was an error as we use `eslint-disable-next-line`
for the `default` case in some switch statements, where the absence of a
warning would mean the switch is no longer exhaustive.