fix: use `describe` instead of `openapi({description})`
Descriptions for query parameters would not be displayed. This turned
out to be because the `description` field was output in the wrong
location.

Our output:

```json
"parameters": [{
  "schema": { "description": "...", ... },
  ...
}]
```

expected output:

```json
"parameters": [{
  "schema": { ... },
  "description": "...",
  ...
}]
```

While researching this, I found that a very similar issue was recently
fixed when using the `.describe(` function. I was not aware of this
function but migrating to it does solve the issue.

See also: https://github.com/asteasolutions/zod-to-openapi/issues/177