test(integration): configure devcontainer and testcontainer to match production environment
While working on listing profiles, I found that `ORDER BY` statements
would not produce results in alphabetical order (as judged by native
swedish speakers). After a lot of research, some IRC chatting and a bug
report or two, I was finally able to configure our production
environment to use a sv_SE ICU collation in the database "dansdata".

This commit configures our local setup for devcontainer and integration
tests to be as similar as possible to this production setup.

Note that `@testcontainers/postgresql` does not support custom
dockerfiles out-of-the-box but I was able to hack around this by looking
at their sources on GitHub