feat(cache): try to optimize url for cache hits on GET/HEAD request
HTTP cache is typically based on the URL. However, this means that two
seemingly equvialent requests for `/example.com?a=1&b=2` and
`/example.com?b=2&a=1` do not share cache! A way to fix this is to
force clients to sort and normalize their query parameters before making
a request. However, asking nicely is probably not going to work so it's
better to have the server do the work.

This commit implements a feature where query parameters are verified to
be sorted and geocoordinates are limited to 3 decimals of precision
(~110 m) before executing the related endpoint's handler.
If the parameters are not sorted or normalized, a 301 redirect is
returned.