fix(15002fd)!: add views for root and leaf nodes + rename venue_parents -> venue_tree
These views will be useful in the listing of venues. The new
`venue_tree` naming also provides a much clearer description of the
view's content.

Note that this change makes changes to _already commited database
migrations_! I'm allowing this as we do not yet have any data in
production, meaning it is still easy to start over from scratch.