test(testcontainers): print testcontainer logs on error
When something goes wrong in the database layer, it can sometimes be
hard to diagnose without connecting to the database or viewing the logs.
Unfortunately, the testcontainer tends to be shut down and removed
before these actions can be taken.

This commit therefore attempts to make things a little easier by
printing the container logs if an error is caught. The commit also adds
a print of the connection uri, making it easier to connect to the
database while stopped at a breakpoint in a test.