test: do not generate fake references in `generateXModel` helpers
Originally, I thought this could be helpful. However, in practice I
found that I usually wanted these values set to `null`/equivalent or
a value created as part of the tests. By changing the generator helpers
to not create random values for fields that are expected to reference
other models, I'm able to reduce the number of lines needed in test
files.