fix: remove unused profile listing methods
These are unused and will be implemented using a different pattern with
`${ProfileType}FilterModel`.

The reason for implementing them differently is that when we have
separate methods for different types of filtering, we cannot apply
multiple filters and sorting algorithms at once. At least not without
taking a big performance hit.

By combining the methods into one, we can optimize our database queries
and perform a lot of the filtering and sorting on the database side.