feat: support free-text search for individuals and organizations
Since venues uses postgis which is not supported by prisma, I had to
implement the listing feature for venues manually in SQL. After doing so,
I found it easy to add free text search when working directly in SQL. I
was also pretty happy with the structure of the query and did not find
it too hard to work with. I therefore decided that (i) if venues support
free text search, so should other profiles and (ii) it's probably better
to implement the listing for individuals and organizations manually in
SQL as well.

Note that this commit also changes the filtering behavior when supplying
an organization or member id. Originally, supplying two organization ids
when listing individuals would return the members of both organizations
(union). This is now changed so that such a query would only return the
individuals who are members of _both_ organizations (intersection). I
made this change after giving the behavior a bit of thought and
concluding that you can still run separated queries to get the members
of both organizations but would have a harder time querying for the
intersection only.

Finally, I don't think the tests for "can limit results by tag" are well
written enough. They should probably use a more controlled setup but I'm
leaving them as is for now.